Change shebang to use bash in marktex
The script uses a string substitution technique in line 3 that is supported by bash, but not by other shell interpreters.
Since e.g. on Debian Linux, the shebang #!/bin/bash points to a different interpreter which does NOT understand the string substitution, executing marktex leads to the error Bad substitution . Another case of this problem is dicussed in this thread on StackOverflow: https://stackoverflow.com/questions/20615217/bash-bad-substitution .

A shebang that explicitly requests bash solves this problem.